refactor - Change return type of interest calculation methods from int to BigDecimal across investment classes

+++ src/main/java/co/invest72/financial_product/presentation/dto/response/FinancialProductSummaryResponse.java
@@ -39,7 +39,7 @@ public class FinancialProductSummaryResponse {
 			.startDate(product.getStartDate())
 			.expirationDate(product.getExpirationDate())
 			.balance(product.getBalanceByLocalDate(today))
-			.expectedInterest(BigDecimal.valueOf(investment.getTotalInterest()))
+			.expectedInterest(investment.getTotalInterest())
 			.progress(product.getProgressByLocalDate(today))
 			.remainingDays(product.getRemainingDaysByLocalDate(today))
 			.createAt(product.getCreatedAt())
