+++ src/main/java/co/invest72/investment/presentation/response/CalculateInvestmentResponse.java
@@ -11,17 +11,17 @@ import lombok.Getter;
 @Getter
 public class CalculateInvestmentResponse {
 	private final BigDecimal totalInvestment;
-	private final int totalInterest;
-	private final int totalTax;
-	private final int totalProfit;
+	private final BigDecimal totalInterest;
+	private final BigDecimal totalTax;
+	private final BigDecimal totalProfit;
 	private final String taxType;
 	private final String taxPercent;
 	private final List<MonthlyInvestmentResult> monthlyDetails;
 	private final List<YearlyInvestmentResult> yearlyDetails;
 
 	@Builder
-	public CalculateInvestmentResponse(BigDecimal totalInvestment, int totalInterest, int totalTax,
-		int totalProfit, String taxType, String taxPercent, List<MonthlyInvestmentResult> monthlyDetails,
+	public CalculateInvestmentResponse(BigDecimal totalInvestment, BigDecimal totalInterest, BigDecimal totalTax,
+		BigDecimal totalProfit, String taxType, String taxPercent, List<MonthlyInvestmentResult> monthlyDetails,
 		List<YearlyInvestmentResult> yearlyDetails) {
 		this.totalInvestment = totalInvestment;
 		this.totalInterest = totalInterest;
